infra/backend/requirements.txt
awscli-local==0.22.0
boto3==1.37.23
botocore==1.37.23
jmespath==1.0.1
localstack-client==2.7
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
pyyaml
